Refactor and simplify project structure
Refactor and simplify the project structure by consolidating related files and improving organization.

* **Configuration Files**: Add `src/config/experiments.py` to consolidate all experiment configurations from `src/config/experiments`.
* **Model Implementations**: Add `src/models/clustering_modules.py` to merge clustering module implementations from `src/models/clustering_module`. Add `src/models/model_implementations.py` to merge model implementations from `src/models`.
* **Data Processing**: Add `src/data/data_processing.py` to group related data processing scripts from `src/data`.
* **Utility Functions**: Add `src/lib/utilities.py` to group utility functions from `src/lib`. Add `src/lib/logging.py` to group logging functions from `src/lib`. Add `src/lib/evaluation.py` to group evaluation functions from `src/lib`.
* **Training Script**: Refactor `src/train.py` to separate concerns and improve readability. Update imports to reflect the new file structure and replace `pytorch_lightning` related code with equivalent PyTorch code.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/liu2048/DeepMVC?shareId=XXXX-XXXX-XXXX-XXXX).
