+++ src/lib/logging.py
@@ -6,7 +6,6 @@ from contextlib import contextmanager
 from copy import deepcopy
 from tabulate import tabulate
 from typing import Union, Dict, Optional, Any
-from pytorch_lightning.loggers.base import LightningLoggerBase
 
 import helpers
 import config
@@ -22,9 +21,8 @@ def fix_cmat(metrics, set_type, cmat_hook=lambda x: x):
     metrics[f"{set_type}_metrics/cmat"] = cmat_hook(cmat)
 
 
-class ConsoleLogger(LightningLoggerBase):
+class ConsoleLogger:
     def __init__(self, ename, ignore_keys=tuple(), print_cmat=True):
-        super(ConsoleLogger, self).__init__()
         self.ignore_keys = list(ignore_keys)
         self.print_cmat = print_cmat
         self.epoch_offset = 0
@@ -34,10 +32,6 @@ class ConsoleLogger(LightningLoggerBase):
         if not self.print_cmat:
             self.ignore_keys.append("val_metrics/cmat")
             self.ignore_keys.append("test_metrics/cmat")
-
-    @property
-    def experiment(self):
-        return None
 
     def log_metrics(self, metrics, step=None):
         print_logs = deepcopy(metrics)
@@ -62,22 +56,9 @@ class ConsoleLogger(LightningLoggerBase):
 
         print(tabulate([values], headers=headers), "\n")
 
-    def log_hyperparams(self, params):
-        pass
 
-    @property
-    def name(self):
-        return self._ename
-
-    @property
-    def version(self):
-        return self._version
-
-
-class WeightsAndBiasesLogger(LightningLoggerBase):
+class WeightsAndBiasesLogger:
     def __init__(self, name, tag, run, cfg, net):
-        super(WeightsAndBiasesLogger, self).__init__()
-
         run_info = get_default_run_info(name, tag, run, cfg)
         self.group = run_info.group
         self.wanbd_run = wandb.init(
@@ -103,10 +84,6 @@ class WeightsAndBiasesLogger(LightningLoggerBase):
         finally:
             self.logging_enabled = True
 
-    @property
-    def experiment(self) -> Any:
-        return self.wanbd_run
-
     @staticmethod
     def _cmat_to_heatmap(cmat):
         ax_labels = list(range(cmat.shape[0]))
@@ -122,17 +99,6 @@ class WeightsAndBiasesLogger(LightningLoggerBase):
         fix_cmat(log_metrics, set_type="test", cmat_hook=self._cmat_to_heatmap)
         self.wanbd_run.log(log_metrics, step=(self.epoch_offset + log_metrics["epoch"]))
 
-    def log_hyperparams(self, params):
-        pass
-
-    @property
-    def name(self):
-        return self.group
-
-    @property
-    def version(self):
-        return self._version
-
     def log_summary(self, val_results, test_results):
         if not self.logging_enabled:
             return
@@ -142,4 +108,3 @@ class WeightsAndBiasesLogger(LightningLoggerBase):
             fix_cmat(logs, set_type=set_type)
             for key, value in logs.items():
                 self.wanbd_run.summary[f"summary/{key}"] = value
-
